refactor model_utils merged with utils
